requirements.txt
fastapi
uvicorn
PyPDF2
reportlab
pdf2docx
docx2pdf
ocrmypdf
pillow
pdf2image
pdfkit
